pyproject.toml
[tool.poetry]
name = "openapi_client_generator"
version = "0.0.1"
description = ""
authors = ["sasano8 <y-sasahara@ys-method.com>"]
license = "MIT"
readme = "README.md"
# repository = "https://github.com/sasano8/asy"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.3.2"
pyyaml = "^5.4.1"


[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9.0"
mypy = "^0.812"
pytest = "^6.2.2"
pre-commit = "^2.12.0"
devtools = "^0.6.1"
fastapi = "^0.65.1"
uvicorn = "^0.14.0"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
